+++ app/src/main/java/org/sopt/dosopttemplate/features/DoAndroidFragment.kt
@@ -1,8 +1,13 @@
 package org.sopt.dosopttemplate.features
 
 import com.example.core_ui.base.BindingFragment
+import dagger.hilt.android.AndroidEntryPoint
 import org.sopt.dosopttemplate.R
 import org.sopt.dosopttemplate.databinding.FragmentDoAndroidBinding
 
+@AndroidEntryPoint
 class DoAndroidFragment : BindingFragment<FragmentDoAndroidBinding>(R.layout.fragment_do_android) {
+    override fun initView() {
+
+    }
 }
