+++ app/src/main/java/org/sopt/dosopttemplate/features/HomeFragment.kt
@@ -1,8 +1,13 @@
 package org.sopt.dosopttemplate.features
 
 import com.example.core_ui.base.BindingFragment
+import dagger.hilt.android.AndroidEntryPoint
 import org.sopt.dosopttemplate.R
 import org.sopt.dosopttemplate.databinding.FragmentHomeBinding
 
+@AndroidEntryPoint
 class HomeFragment : BindingFragment<FragmentHomeBinding>(R.layout.fragment_home) {
+    override fun initView() {
+
+    }
 }
