+++ app/src/main/res/values/strings.xml
@@ -26,4 +26,7 @@
     <string name="content_description_iv_main_profile">프로필 사진입니다.</string>
     <string name="btn_main_sign_out_label">SIGN OUT</string>
     <string name="message_back_to_exit_pressed">뒤로 가기 버튼을 한 번 더 누르면 종료됩니다.</string>
+    <string name="menu_do_android_label">DO ANDROID</string>
+    <string name="menu_home_label">홈</string>
+    <string name="menu_my_page_label">마이페이지</string>
 </resources>
